Bug Fix
Removed the String() because without it the code still works as
intended.

The Zeros were appending to each other when pressed multiple times.
'000000'
To avoid it, when we used to check for 'value === 0', I've changed
it to "value == '0'", so it checks if value is equal to String(0) or
Integer(0) at the same time.